Explicit 'test' task; extended slow tests
