+++ test/piezo.js
@@ -226,20 +226,6 @@ exports["Piezo"] = {
     test.done();
   },
 
-  toneStopsAfterTime: function(test) {
-    test.expect(2);
-
-    this.piezo.tone(1915, 10);
-    var timerSpy = sinon.spy(this.piezo.timer, "clearInterval");
-
-    setTimeout(function() {
-      test.ok(timerSpy.called);
-      test.equal(this.piezo.timer, undefined);
-
-      test.done();
-    }.bind(this), 30);
-  },
-
   toneWhileNewToneIsPlayingCancelsExisting: function(test) {
     test.expect(1);
 
@@ -395,53 +381,6 @@ exports["Piezo"] = {
     });
   },
 
-  playSingleNoteTune: function(test) {
-    var tempo = 10000;
-    test.expect(2);
-    var freqSpy = sinon.spy(this.piezo, "frequency");
-    var returned = this.piezo.play({
-      song: "c4",
-      tempo: tempo // Make it real fast
-    });
-    setTimeout(function() {
-      test.ok(freqSpy.calledOnce);
-      test.ok(freqSpy.calledWith(Piezo.Notes["c4"], 60000 / tempo));
-      test.done();
-    }.bind(this), 10);
-  },
-
-  playSingleNoteTuneNoOctave: function(test) {
-    var tempo = 10000;
-    test.expect(2);
-    var freqSpy = sinon.spy(this.piezo, "frequency");
-    var returned = this.piezo.play({
-      song: "c#",
-      tempo: tempo // Make it real fast
-    });
-    setTimeout(function() {
-      test.ok(freqSpy.calledOnce);
-      test.ok(freqSpy.calledWith(Piezo.Notes["c#4"], 60000 / tempo));
-      test.done();
-    }.bind(this), 10);
-  },
-
-  playSongWithCallback: function(test) {
-    var tempo = 10000,
-      myCallback = sinon.spy(),
-      tune = {
-        song: ["c4"],
-        tempo: tempo
-      };
-    test.expect(2);
-
-    var returned = this.piezo.play(tune, myCallback);
-    setTimeout(function() {
-      test.ok(myCallback.calledOnce);
-      test.ok(myCallback.calledWith(tune));
-      test.done();
-    }.bind(this), 10);
-  },
-
   playCanDealWithWonkyValues: function(test) {
     var tempo = 10000,
       tune = {
@@ -460,21 +399,4 @@ exports["Piezo"] = {
     }.bind(this));
 
   },
-  /**
-   * This is a slow test by necessity because the default tempo (which can"t
-   * be overridden if `play` is invoked with a non-object arg) is 250.
-   * It does pass.
-   */
-  /*
-  playSingleNote: function(test) {
-    test.expect(2);
-    var freqSpy = sinon.spy(this.piezo, "frequency");
-    var returned = this.piezo.play("c4");
-    setTimeout(function() {
-      test.ok(freqSpy.calledOnce);
-      test.ok(freqSpy.calledWith(Piezo.Notes["c4"], (60000 / 250)));
-      test.done();
-    }.bind(this), 260);
-  }
-  */
 };
